WIP: Merge edit and view activities into one. needs testing. should probably figure out a better way to handle tags
