+++ RecipeBinderApp/src/main/AndroidManifest.xml
@@ -23,14 +23,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activities.EditRecipeActivity"
-            android:label="@string/label_edit_activity"
-            android:theme="@style/AppTheme.NoActionBar"
-            android:windowSoftInputMode="adjustPan" />
-        <activity
             android:name=".activities.ViewRecipeActivity"
             android:label="Recipe Viewer"
-            android:theme="@style/AppTheme.NoActionBar" />
+            android:theme="@style/AppTheme.NoActionBar"
+            android:windowSoftInputMode="adjustPan" />
     </application>
 
 </manifest>
