+++ RecipeBinderApp/src/main/res/drawable-v21/edit_to_save.xml
@@ -8,21 +8,29 @@
             android:viewportWidth="24"
             android:viewportHeight="24">
             <group
-                android:name="group"
+                android:name="pencil_group"
                 android:pivotX="12"
                 android:pivotY="12">
                 <path
-                    android:name="path"
+                    android:name="pencil"
                     android:fillColor="#FF000000"
                     android:pathData="M 3 17.25 L 3 21 L 6.75 21 L 17.81 9.94 L 14.06 6.19 L 3 17.25 Z M 20.71 7.04 C 21.1 6.65 21.1 6.02 20.71 5.63 L 18.37 3.29 C 17.98 2.9 17.35 2.9 16.96 3.29 L 15.13 5.12 L 18.88 8.87 L 20.71 7.04 Z" />
             </group>
+            <group android:name="save_group">
+                <path
+                    android:name="save"
+                    android:fillAlpha="0"
+                    android:fillColor="#FF000000"
+                    android:pathData="M 17 3 L 5 3 C 3.89 3 3 3.9 3 5 L 3 19 C 3 20.1 3.89 21 5 21 L 19 21 C 20.1 21 21 20.1 21 19 L 21 7 L 17 3 Z M 12 19 C 10.34 19 9 17.66 9 16 C 9 14.34 10.34 13 12 13 C 13.66 13 15 14.34 15 16 C 15 17.66 13.66 19 12 19 Z M 15 9 L 5 9 L 5 5 L 15 5 L 15 9 Z" />
+            </group>
         </vector>
     </aapt:attr>
-    <target android:name="group">
+    <target android:name="pencil_group">
         <aapt:attr name="android:animation">
             <set>
                 <objectAnimator
                     android:duration="300"
+                    android:interpolator="@android:anim/linear_interpolator"
                     android:propertyName="rotation"
                     android:valueFrom="0"
                     android:valueTo="180"
@@ -44,6 +52,18 @@
                     android:valueTo="-14"
                     android:valueType="floatType" />
             </set>
+        </aapt:attr>
+    </target>
+    <target android:name="save">
+        <aapt:attr name="android:animation">
+            <objectAnimator
+                android:duration="90"
+                android:interpolator="@android:interpolator/linear_out_slow_in"
+                android:propertyName="fillAlpha"
+                android:startOffset="410"
+                android:valueFrom="0"
+                android:valueTo="1"
+                android:valueType="floatType" />
         </aapt:attr>
     </target>
 </animated-vector>
