+++ RecipeBinderApp/src/main/res/drawable-v21/save_to_edit.xml
@@ -1,0 +1,39 @@
+<animated-vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt">
+    <aapt:attr name="android:drawable">
+        <vector
+            android:name="vector"
+            android:width="24dp"
+            android:height="24dp"
+            android:viewportWidth="24"
+            android:viewportHeight="24">
+            <group android:name="save_group">
+                <path
+                    android:name="save"
+                    android:fillColor="#FF000000"
+                    android:pathData="M 17 3 L 5 3 C 3.89 3 3 3.9 3 5 L 3 19 C 3 20.1 3.89 21 5 21 L 19 21 C 20.1 21 21 20.1 21 19 L 21 7 L 17 3 Z M 12 19 C 10.34 19 9 17.66 9 16 C 9 14.34 10.34 13 12 13 C 13.66 13 15 14.34 15 16 C 15 17.66 13.66 19 12 19 Z M 15 9 L 5 9 L 5 5 L 15 5 L 15 9 Z" />
+            </group>
+        </vector>
+    </aapt:attr>
+    <target android:name="save">
+        <aapt:attr name="android:animation">
+            <set>
+                <objectAnimator
+                    android:duration="200"
+                    android:interpolator="@android:anim/linear_interpolator"
+                    android:propertyName="pathData"
+                    android:valueFrom="M 17 3 L 5 3 C 3.89 3 3 3.9 3 5 L 3 19 C 3 20.1 3.89 21 5 21 L 19 21 C 20.1 21 21 20.1 21 19 L 21 7 L 17 3 Z M 12 19 C 10.34 19 9 17.66 9 16 C 9 14.34 10.34 13 12 13 C 13.66 13 15 14.34 15 16 C 15 17.66 13.66 19 12 19 L 12 19 M 15 9 L 5 9 L 5 5 L 15 5 L 15 9 L 15 9"
+                    android:valueTo="M 17 3 L 5 3 C 3.89 3 3 3.9 3 5 L 3 19 C 3 20.1 3.89 21 5 21 L 19 21 C 20.1 21 21 20.1 21 19 L 21 7 L 17 3 Z M 12 16 C 12 16 12 16 12 16 C 12 16 12 16 12 16 C 12 16 12 16 12 16 C 12 16 12 16 12 16 L 12 16 M 10 7 L 10 7 L 10 7 L 10 7 L 10 7 L 10 7"
+                    android:valueType="pathType" />
+                <objectAnimator
+                    android:duration="300"
+                    android:interpolator="@android:interpolator/fast_out_slow_in"
+                    android:propertyName="pathData"
+                    android:startOffset="200"
+                    android:valueFrom="M 14.005 3 L 21 9.838 L 21 7.791 L 21 7 C 19.667 5.667 18.333 4.333 17 3 L 17 3 C 16.757 3 16.515 3 16.272 3 L 14.005 3 M 21 9.838 L 14.005 3 L 5 3 C 3.89 3 3 3.9 3 5 L 3 19 C 3 20.1 3.89 21 5 21 L 19 21 C 20.1 21 21 20.1 21 19 L 21 9.838 M 12 16 L 12 16 L 12 16 L 12 16 L 12 16 L 12 16 M 10 7 L 10 7 L 10 7 L 10 7 L 10 7 L 10 7"
+                    android:valueTo="M 15.13 5.12 L 18.88 8.87 L 20.71 7.04 L 20.71 7.04 C 21.1 6.65 21.1 6.02 20.71 5.63 L 18.37 3.29 C 17.98 2.9 17.35 2.9 16.96 3.29 L 15.13 5.12 M 17.81 9.94 L 14.06 6.19 L 4 16.25 C 3.666 16.584 3.333 16.917 3 17.25 L 3 17.25 C 3 18.5 3 19.75 3 21 L 3.895 21 C 4.847 21 5.798 21 6.75 21 L 17.81 9.94 M 12 16 L 12 16 L 12 16 L 12 16 L 12 16 L 12 16 M 10 7 L 10 7 L 10 7 L 10 7 L 10 7 L 10 7"
+                    android:valueType="pathType" />
+            </set>
+        </aapt:attr>
+    </target>
+</animated-vector>
