+++ RecipeBinderApp/src/main/res/values/dimens.xml
@@ -1,7 +1,5 @@
 <resources>
     <dimen name="fab_margin">16dp</dimen>
-    <dimen name="padding_recipe_view">10dp</dimen>
-    <dimen name="title_text_size">18pt</dimen>
     <dimen name="info_text_size">14pt</dimen>
     <dimen name="dialog_margin">4dp</dimen>
     <dimen name="ingredient_list_text_size">10pt</dimen>
